Lift logic up to case timeline component
