+++ client/app/queue/components/EditNODModal.jsx
@@ -1,27 +1,12 @@
 import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
 import ReactMarkdown from 'react-markdown';
 import PropTypes from 'prop-types';
 import Modal from 'app/components/Modal';
 import DateSelector from 'app/components/DateSelector';
 import COPY from 'app/../COPY';
-import { showSuccessMessage, resetSuccessMessages } from '../uiReducer/uiActions';
 
-export const EditNODModal = ({ onCancel, nodDate }) => {
+export const EditNODModal = ({ onCancel, onSubmit, nodDate }) => {
   const [receiptDate, setReceiptDate] = useState(nodDate);
-  const dispatch = useDispatch();
-
-  const onSubmit = () => {
-
-    const successMessage = {
-      title: COPY.EDIT_NOD_DATE_SUCCESS_ALERT_TITLE,
-      detail: COPY.EDIT_NOD_DATE_SUCCESS_ALERT_MESSAGE,
-    };
-
-    dispatch(showSuccessMessage(successMessage));
-    setTimeout(() => dispatch(resetSuccessMessages()), 5000);
-    onCancel();
-  };
 
   const buttons = [
     {
@@ -32,7 +17,7 @@ export const EditNODModal = ({ onCancel, nodDate }) => {
     {
       classNames: ['usa-button', 'usa-button-primary'],
       name: 'Submit',
-      onClick: onSubmit
+      onClick: () => onSubmit()
     }
   ];
 
@@ -42,6 +27,7 @@ export const EditNODModal = ({ onCancel, nodDate }) => {
     <Modal
       title={COPY.EDIT_NOD_DATE_MODAL_TITLE}
       onCancel={onCancel}
+      onSubmit={onSubmit}
       closeHandler={onCancel}
       buttons={buttons}>
       <div>
@@ -60,5 +46,6 @@ export const EditNODModal = ({ onCancel, nodDate }) => {
 
 EditNODModal.propTypes = {
   onCancel: PropTypes.func.isRequired,
+  onSubmit: PropTypes.func.isRequired,
   nodDate: PropTypes.string.isRequired
 };
